refactor(persistence): rename repos to Postgres* and move config files
