+++ src/main/kotlin/io/sdkman/state/Application.kt
@@ -2,13 +2,13 @@ package io.sdkman.state
 
 import io.ktor.server.application.*
 import io.sdkman.state.adapter.primary.rest.*
-import io.sdkman.state.adapter.secondary.persistence.AuditRepositoryImpl
-import io.sdkman.state.adapter.secondary.persistence.HealthRepositoryImpl
-import io.sdkman.state.adapter.secondary.persistence.TagsRepositoryImpl
-import io.sdkman.state.adapter.secondary.persistence.VersionsRepository
+import io.sdkman.state.adapter.secondary.persistence.PostgresAuditRepository
+import io.sdkman.state.adapter.secondary.persistence.PostgresHealthRepository
+import io.sdkman.state.adapter.secondary.persistence.PostgresTagRepository
+import io.sdkman.state.adapter.secondary.persistence.PostgresVersionRepository
 import io.sdkman.state.application.service.TagServiceImpl
 import io.sdkman.state.application.service.VersionServiceImpl
-import io.sdkman.state.config.configureAppConfig
+import io.sdkman.state.config.*
 import io.sdkman.state.plugins.*
 
 fun main(args: Array<String>) =
@@ -25,13 +25,13 @@ fun Application.module() {
     configureSerialization()
     configureBasicAuthentication(appConfig.apiAuthenticationConfig)
 
-    val versionsRepo = VersionsRepository()
-    val tagsRepo = TagsRepositoryImpl()
-    val auditRepo = AuditRepositoryImpl()
+    val versionsRepo = PostgresVersionRepository()
+    val tagsRepo = PostgresTagRepository()
+    val auditRepo = PostgresAuditRepository()
 
     configureRouting(
         versionService = VersionServiceImpl(versionsRepo, tagsRepo, auditRepo),
         tagService = TagServiceImpl(tagsRepo, auditRepo),
-        healthRepo = HealthRepositoryImpl(),
+        healthRepo = PostgresHealthRepository(),
     )
 }
