+++ src/main/kotlin/io/sdkman/state/config/Migration.kt
@@ -1,7 +1,6 @@
-package io.sdkman.state.plugins
+package io.sdkman.state.config
 
 import io.ktor.server.application.*
-import io.sdkman.state.config.DatabaseConfig
 import org.flywaydb.core.Flyway
 
 fun Application.configureDatabaseMigration(config: DatabaseConfig) {
