+++ src/test/kotlin/io/sdkman/state/support/Application.kt
@@ -3,10 +3,10 @@ package io.sdkman.state.support
 import io.ktor.server.config.*
 import io.ktor.server.testing.*
 import io.sdkman.state.adapter.primary.rest.configureRouting
-import io.sdkman.state.adapter.secondary.persistence.AuditRepositoryImpl
-import io.sdkman.state.adapter.secondary.persistence.HealthRepositoryImpl
-import io.sdkman.state.adapter.secondary.persistence.TagsRepositoryImpl
-import io.sdkman.state.adapter.secondary.persistence.VersionsRepository
+import io.sdkman.state.adapter.secondary.persistence.PostgresAuditRepository
+import io.sdkman.state.adapter.secondary.persistence.PostgresHealthRepository
+import io.sdkman.state.adapter.secondary.persistence.PostgresTagRepository
+import io.sdkman.state.adapter.secondary.persistence.PostgresVersionRepository
 import io.sdkman.state.application.service.TagServiceImpl
 import io.sdkman.state.application.service.VersionServiceImpl
 import io.sdkman.state.config.configureAppConfig
@@ -21,14 +21,14 @@ fun withTestApplication(fn: suspend (ApplicationTestBuilder.() -> Unit)) {
             val dbConfig = configureAppConfig(environment).databaseConfig
             configureDatabase(dbConfig)
 
-            val versionsRepo = VersionsRepository()
-            val tagsRepo = TagsRepositoryImpl()
-            val auditRepo = AuditRepositoryImpl()
+            val versionsRepo = PostgresVersionRepository()
+            val tagsRepo = PostgresTagRepository()
+            val auditRepo = PostgresAuditRepository()
 
             configureRouting(
                 versionService = VersionServiceImpl(versionsRepo, tagsRepo, auditRepo),
                 tagService = TagServiceImpl(tagsRepo, auditRepo),
-                healthRepo = HealthRepositoryImpl(),
+                healthRepo = PostgresHealthRepository(),
             )
         }
         fn(this)
